+++ point_cloud_io/operators/__init__.py
@@ -4,11 +4,13 @@ from .export_e57 import EXPORT_OT_e57, menu_func_export_e57
 from .export_las import EXPORT_OT_las, menu_func_export_las
 from .export_pcd import EXPORT_OT_pcd, menu_func_export_pcd
 from .export_ply import EXPORT_OT_ply, menu_func_export_ply
+from .export_pts import EXPORT_OT_pts, menu_func_export_pts
 from .export_xyz import EXPORT_OT_xyz, menu_func_export_xyz
 from .import_e57 import IMPORT_OT_e57, menu_func_import_e57
 from .import_las import IMPORT_OT_las, menu_func_import_las
 from .import_pcd import IMPORT_OT_pcd, menu_func_import_pcd
 from .import_ply import IMPORT_OT_ply, menu_func_import_ply
+from .import_pts import IMPORT_OT_pts, menu_func_import_pts
 from .import_xyz import IMPORT_OT_xyz, menu_func_import_xyz
 
 _classes = (
@@ -22,6 +24,8 @@ _classes = (
     EXPORT_OT_pcd,
     IMPORT_OT_xyz,
     EXPORT_OT_xyz,
+    IMPORT_OT_pts,
+    EXPORT_OT_pts,
 )
 _import_menu_funcs = (
     menu_func_import_e57,
@@ -29,6 +33,7 @@ _import_menu_funcs = (
     menu_func_import_las,
     menu_func_import_pcd,
     menu_func_import_xyz,
+    menu_func_import_pts,
 )
 _export_menu_funcs = (
     menu_func_export_e57,
@@ -36,6 +41,7 @@ _export_menu_funcs = (
     menu_func_export_las,
     menu_func_export_pcd,
     menu_func_export_xyz,
+    menu_func_export_pts,
 )